Moved column render to utils
