+++ app/javascript/src/components/Dashboard/Contacts/Table/index.jsx
@@ -2,14 +2,15 @@ import React from "react";
 
 import { Table as NeetoTable, Pagination } from "neetoui";
 
-import { COLUMN_HEADERS, CONTACTS } from "./constants";
+import { CONTACTS } from "./constants";
+import { buildColumns } from "./utils";
 
 const Table = ({ setIsDeleteOpen }) => (
   <>
     <div className="w-full">
       <NeetoTable
         className="odd:bg-gray-100"
-        columnData={COLUMN_HEADERS({ setIsDeleteOpen })}
+        columnData={buildColumns({ setIsDeleteOpen })}
         rowData={CONTACTS}
       />
     </div>
